+++ src/FSharp.Formatting.ApiDocs/GenerateMarkdown.fs
@@ -622,5 +622,5 @@ type MarkdownRender(model: ApiDocModel, ?menuTemplateFolder: string) =
         for (relativeFile, render) in x.Pages(collectionName) do
             let outFile = Path.Combine(outDir, relativeFile)
             let outputText = render templateOpt globalParameters
-            printfn "  Generating %s" outFile
+            logger.Debugf "  Generating %s" outFile
             SimpleTemplating.WriteOutputFile(outFile, outputText)